CATS pre-start: explicitly add SG for endpoint
The default configuration we use for vagrant boxes uses a private
network address (192.168.77.77) for the external endpoint for the CF
cluster.  This means that it is not covered by the public networks
default security group, and therefore apps won't be allowed to connect
to it.  This breaks the route services test where it attempts to connect
to an exposed port from within a different application.