HHM178x6 - Bump cf cli to 6.33.1, as requested by the new :cat:
ATTENTION: Switched to SUSE-specific bucket `s3://scf-helper-release-blobs`, from `s3://cf-release-blobs`.
It was needed because cli 6.33.1 could not be found in the bucket.
Using our own bucket we have more control over the contents.
More work too.

The cats are running with the change, although the outcome re success/failure is not known yet.